pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-camp-note"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jpype1>=1.6.0",
    "jupyter>=1.1.1",
    "konlpy>=0.6.0",
    "lyricsgenius>=3.11.0",
    "mysql-connector-python>=9.6.0",
    "python-dotenv>=1.2.2",
    "seaborn>=0.13.2",
    "streamlit>=1.56.0",
    "wordcloud>=1.9.6",
    "pydantic>=2.0.0",
    "pandera[pandas]>=0.13.0",
    "xmltodict>=0.13.0",
    "folium>=0.20.0",
]

[tool.setuptools]
packages = ["week1", "week2", "week3"]
py-modules = ["config"]
